+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.11.1-dev0
+## 0.11.1-dev1
 
 ### Enhancements
 
@@ -7,6 +7,7 @@
 ### Fixes
 
 * **Do not extract text of `<style>` tags in HTML.** `<style>` tags containing CSS in invalid positions previously contributed to element text. Do not consider text node of a `<style>` element as textual content.
+* **Fix DOCX merged table cell repeats cell text.** Only include text for a merged cell, not for each underlying cell spanned by the merge.
 
 ## 0.11.0